+++ CHANGELOG.md
@@ -12,7 +12,8 @@
 - Update BTP Scanner,
 - Update CAP Scanner,
 - Update NPM Scanner,
-- Update Secret Scanner.
+- Update Secret Scanner,
+- Update Documentation.
 
 ## v1.2.2 (2026-06-21)
 - Security and Quality - Uncontrolled data used in path expression - utils/fileParser.js:104,